+++ integration_tests/src/tests/mod.rs
@@ -28,7 +28,8 @@ mod replica_update_test;
 pub use replica_update_test::*;
 mod terminate_test;
 pub use terminate_test::*;
-// mod verified_claim_test;
+mod verified_claim_test;
+pub use verified_claim_test::*;
 // mod verifreg_remove_datacap_test;
 mod withdraw_balance_test;
 pub use withdraw_balance_test::*;
